requirements.txt
fastapi[all]
uvicorn[standard]

cryptography
pyjwt